added use VAT inclusion settings in price product

+++ tests/bin/before_script.sh
@@ -9,7 +9,12 @@ PRESTASHOP_DIR=$TRAVIS_BUILD_DIR/../PrestaShop
 cd $PRESTASHOP_DIR
 
 if [ -z $BRANCH ]; then
-    cp tests/parameters.yml.travis app/config/parameters.yml
+	if [ -f "tests/parameters.yml.travis" ]; then
+	    cp tests/parameters.yml.travis app/config/parameters.yml
+    else
+        cp tests-legacy/parameters.yml.travis app/config/parameters.yml
+    fi
+
     bash travis-scripts/install-prestashop
 else
     bash travis-scripts/install-prestashop.sh
